Move dependendecies meta data into code/
