+++ modelstore/metadata/metadata.py
@@ -14,8 +14,8 @@
 from datetime import datetime
 
 import modelstore
-from modelstore.metadata import dependencies, runtime
-from modelstore.metadata.code import revision
+from modelstore.metadata import runtime
+from modelstore.metadata.code import revision, dependencies
 
 
 def generate_for_model(
