+++ modelstore/models/model_manager.py
@@ -20,7 +20,7 @@ from abc import ABC, ABCMeta, abstractmethod
 
 import numpy as np
 from modelstore.metadata import metadata
-from modelstore.metadata.dependencies import save_dependencies, save_model_info
+from modelstore.metadata.code.dependencies import save_dependencies, save_model_info
 from modelstore.storage.storage import CloudStorage
